+++ client/app/service/bean_factory.js
@@ -1,0 +1,4 @@
+export function beanFactory() {
+  var bean = {};
+  return bean;
+}
